+++ priv/alphonse/lib/atom.js
@@ -1,28 +1,25 @@
 let Atom;
 
-Atom = function(value){
-  if (!(this instanceof Atom)){
-    return new Atom(value);
-  }
-
-  this.value = value;
-};
-
-Atom.prototype.toString = function () {
-  return this.value;
+Atom = function(_value){
+  return Symbol.for(_value);
 };
 
 Atom.__MODULE__ = Atom("Atom");
 
 Atom.to_string = function (atom) {
-  return atom.toString();
+  let atomString = atom.toString();
+  let indexOfOpenParen = atomString.indexOf("(");
+  let indexOfCloseParen = atomString.lastIndexOf(")");
+  return atomString.substring(indexOfOpenParen + 1, indexOfCloseParen);
 };
 
 Atom.to_char_list = function (atom) {
   let char_list = [];
 
-  for(let i = 0; i < atom.value.length; i++){
-    char_list.push(atom.value.charAt(i));
+  let atomString = Atom.to_string(atom);
+
+  for(let i = 0; i < atomString.length; i++){
+    char_list.push(atomString.charAt(i));
   }
 
   return char_list;