+++ priv/alphonse/lib/enum.js
@@ -1,4 +1,5 @@
 import Tuple from './tuple';
+import List from './list';
 import Atom from './atom';
 import Kernel from './kernel';
 
@@ -22,9 +23,9 @@ let Enum = {
   },
 
   at: function(collection, n, the_default = null){
-    for (var i = 0; i < collection.length; i++) {
+    for (var i = 0; i < collection.length(); i++) {
       if(i === n){
-        return collection[i];
+        return collection.get(i);
       }
     }
 
@@ -35,12 +36,12 @@ let Enum = {
     if(fun == null){
       return Kernel.length(collection);
     }else{
-      return Kernel.length(collection.filter(fun));
+      return Kernel.length(collection.value().filter(fun));
     }
   },
 
   each: function(collection, fun){
-    [].forEach.call(collection, fun);
+    [].forEach.call(collection.value(), fun);
   },
 
   empty__qmark__: function(collection){
@@ -48,28 +49,28 @@ let Enum = {
   },
 
   filter: function(collection, fun){
-    [].filter.call(collection, fun);
+    [].filter.call(collection.value(), fun);
   },
 
   map: function(collection, fun){
-    [].map.call(collection, fun);
+    [].map.call(collection.value(), fun);
   },
 
   map_reduce: function(collection, acc, fun){
     let mapped = [];
     let the_acc = acc;
 
-    for (var i = 0; i < collection.length; i++) {
-      let tuple = fun(collection[i], the_acc);
+    for (var i = 0; i < collection.length(); i++) {
+      let tuple = fun(collection.get(i), the_acc);
       the_acc = tuple[1];
       mapped.push(tuple[0]);
     }
 
-    return Tuple(mapped, the_acc);
+    return Tuple(List(...mapped), the_acc);
   },
 
   member: function(collection, value){
-    for(let x of collection){
+    for(let x of collection.value()){
       if(x === value){
         return true;
       }
@@ -81,8 +82,8 @@ let Enum = {
   reduce: function(collection, acc, fun){
     let the_acc = acc;
 
-    for (var i = 0; i < collection.length; i++) {
-      the_acc = fun(collection[i], the_acc);
+    for (var i = 0; i < collection.length(); i++) {
+      the_acc = fun(collection.get(i), the_acc);
     }
 
     return the_acc;