+++ priv/alphonse/lib/mutable.js
@@ -1,4 +1,8 @@
+import Atom from './atom';
+
 let Mutable = {
+  __MODULE__: Atom('Mutable'),
+
   update: function(obj, prop, value){
     obj[prop] = value;
   }