+++ priv/alphonse/lib/range.js
@@ -1,22 +1,40 @@
 import Atom from './atom';
 
-let Range = function(first, last){
+let Range = function(_first, _last){
   if (!(this instanceof Range)){
-    return new Range(first, last);
+    return new Range(_first, _last);
   }
 
-  this.first = first;
-  this.last = last;
-  this.range = [];
+  this.first = function(){
+    return _first;
+  };
 
-  for(let i = first; i <= last; i++){
-    this.range.push(i);
+  this.last = function(){
+    return _last;
+  };
+
+  let _range = [];
+
+  for(let i = _first; i <= _last; i++){
+    _range.push(i);
   }
 
-  this.length = this.range.length;
+  _range = Object.freeze(_range);
+
+  this.value = function(){
+    return _range;
+  };
+
+  this.length = function(){
+    return _range.length;
+  };
 };
 
 Range.__MODULE__ = Atom('Range');
+
+Range.prototype[Symbol.iterator] = function(){
+  return this.value()[Symbol.iterator]();
+};
 
 Range.new = function (first, last) {
   return Range(first, last);