+++ priv/alphonse/lib/tuple.js
@@ -1,27 +1,39 @@
 import Atom from './atom';
+import List from './list';
 
 let Tuple = function(...args){
   if (!(this instanceof Tuple)){
     return new Tuple(...args);
   }
 
-  this.value = args;
-  this.length = args.length;
+  let _value = Object.freeze(args);
 
-  for (var i = 0; i < args.length; i++) {
-    this[i] = args[i];
-  }
+  this.value = function(){
+    return _value;
+  };
+
+  this.length = function(){
+    return _value.length;
+  };
+
+  this.get = function(i){
+    return _value[i];
+  };
 };
 
-Tuple.prototype.__MODULE__ = Atom('Tuple');
+Tuple.__MODULE__ = Atom('Tuple');
+
+Tuple.prototype[Symbol.iterator] = function(){
+  return this.value()[Symbol.iterator]();
+};
 
 Tuple.prototype.toString = function(){
   var i, s = "";
-  for (i = 0; i < this.length; i++) {
+  for (i = 0; i < this.length(); i++) {
     if (s !== "") {
       s += ", ";
     }
-    s += this[i].toString();
+    s += this.get(i).toString();
   }
 
   return "{" + s + "}";
@@ -34,9 +46,9 @@ Tuple.to_string = function(tuple){
 Tuple.delete_at = function(tuple, index){
   let new_list = [];
 
-  for (var i = 0; i < tuple.value.length; i++) {
+  for (var i = 0; i < tuple.length(); i++) {
     if(i !== index){
-      new_list.push(tuple.value[i]);
+      new_list.push(tuple.get(i));
     }
   }
 
@@ -56,13 +68,13 @@ Tuple.duplicate = function(data, size){
 Tuple.insert_at = function(tuple, index, term){
   let new_tuple = [];
 
-  for (var i = 0; i <= tuple.value.length; i++) {
+  for (var i = 0; i <= tuple.length(); i++) {
     if(i === index){
       new_tuple.push(term);
       i++;
-      new_tuple.push(tuple.value[i]);
+      new_tuple.push(tuple.get(i));
     }else{
-      new_tuple.push(tuple.value[i]);
+      new_tuple.push(tuple.get(i));
     }
   }
 
@@ -70,11 +82,17 @@ Tuple.insert_at = function(tuple, index, term){
 };
 
 Tuple.from_list = function(list){
-  return Tuple.apply(null, list);
+  return Tuple.apply(null, list.value());
 };
 
 Tuple.to_list = function(tuple){
-  return tuple.value;
+  let new_list = [];
+
+  for (var i = 0; i < tuple.length(); i++) {
+    new_list.push(tuple.get(i));
+  }
+
+  return List(...new_list);
 };
 
 export default Tuple;