+++ priv/alphonse/tests/range.spec.js
@@ -6,9 +6,9 @@ describe('Range', function(){
   describe('new', function(){
     it('must create a new Range', function(){
       let range = Range.new(0,2);
-      expect(range.first).to.equal(0);
-      expect(range.last).to.equal(2);
-      expect(range.range.length).to.equal(3);
+      expect(range.first()).to.equal(0);
+      expect(range.last()).to.equal(2);
+      expect(range.length()).to.equal(3);
     })
   })
 })