+++ priv/alphonse/tests/tuple.spec.js
@@ -1,6 +1,7 @@
 "use strict"
 
 var Tuple = require('../lib/tuple');
+var List = require('../lib/list');
 var expect = require('chai').expect;
 
 describe('Tuple', function(){
@@ -8,13 +9,13 @@ describe('Tuple', function(){
   describe('duplicate', function(){
     it('must make a tuple with the value duplicated the specified amount of times', function(){
       let t = Tuple.duplicate("value", 3);
-      expect(t.length).to.equal(3);
-      expect(t[0]).to.equal("value");
-      expect(t[1]).to.equal("value");
-      expect(t[2]).to.equal("value");
+      expect(t.length()).to.equal(3);
+      expect(t.get(0)).to.equal("value");
+      expect(t.get(1)).to.equal("value");
+      expect(t.get(2)).to.equal("value");
 
       t = Tuple.duplicate("value", 0);
-      expect(t.length).to.equal(0);
+      expect(t.length()).to.equal(0);
     })
   })
 
@@ -24,7 +25,7 @@ describe('Tuple', function(){
 
       t = Tuple.delete_at(t, 0);
 
-      expect(t[0]).to.equal(2);
+      expect(t.get(0)).to.equal(2);
     })
   })
 
@@ -37,7 +38,7 @@ describe('Tuple', function(){
 
   describe('from_list', function(){
     it('must create a tuple from a list', function(){
-      let list = [1, 2, 3];
+      let list = List(1, 2, 3);
       expect(Tuple.from_list(list).toString()).to.equal("{1, 2, 3}");
     })
   })
@@ -45,7 +46,7 @@ describe('Tuple', function(){
   describe('to_list', function(){
     it('must create a list from a tuple', function(){
       let t = Tuple(1, 2, 3);
-      expect(Tuple.to_list(t).length).to.equal(3);
+      expect(Tuple.to_list(t).length()).to.equal(3);
     })
   })
 })