+++ test/translator/assignment_test.exs
@@ -23,20 +23,16 @@ defmodule ElixirScript.Translator.Assignment.Test do
 
     ex_ast = quote do: {a, b} = {1, 2}
     js_code = """
-        let _ref = Tuple(1, 2);
-        let a = _ref[0];  
-        let b = _ref[1];
+        let [a, b] = Tuple(1, 2);
+        let _ref = Tuple(a, b);
     """
 
     assert_translation(ex_ast, js_code)
 
     ex_ast = quote do: {a, _, c} = {1, 2, 3}
     js_code = """
-        let _ref = Tuple(1, 2, 3);
-
-        let a = _ref[0];  
-        let undefined = _ref[1];  
-        let c = _ref[2];
+        let [a, undefined, c] = Tuple(1, 2, 3);
+        let _ref = Tuple(a, undefined, c);
     """
 
     assert_translation(ex_ast, js_code)
@@ -46,11 +42,11 @@ defmodule ElixirScript.Translator.Assignment.Test do
     js_code = """
         let _ref = Tuple(1, 2, 3);
 
-        if(!Kernel.match__qmark__(a, _ref[0]))
+        if(!Kernel.match__qmark__(a, _ref.get(0)))
           throw new MatchError('no match of right hand side value');
 
-        let undefined = _ref[1];
-        let c = _ref[2];
+        let undefined = _ref.get(1);
+        let c = _ref.get(2);
     """
 
     assert_translation(ex_ast, js_code)